Moved upload_wasm
