+++ cw-orch-daemon/src/senders/mod.rs
@@ -9,7 +9,6 @@ mod cosmos_batch;
 mod cosmos_options;
 mod query_only;
 
-pub(crate) use cosmos::upload_wasm;
 pub use {
     cosmos::{CosmosSender, Wallet},
     cosmos_batch::{options::CosmosBatchOptions, BatchDaemon, CosmosBatchSender},
